Rework BulkRenamer to use IRenameOperations
This reworks the guts of BulkRenamer to instead use an interface. This
will be used to allow us to apply these operations in an arbitrary,
user-defined order. In doing so, we'll be able to fix issues with
unclear and ambiguous order of operations. It also cleans up the code a
bunch.

Fixes Issue #41.

For Posterity sake, the commits for this were made as follows:

Use an IRenameOperation for ReplacementStrings

Convert Add Prefix / Suffix to an IRenameOperation

Add and replace TrimCharactersOperation

Rework Enumerate to an EnumerateOperation

Derive all operations from a BaseOperation and remove shared code

Comply with StyleCop