Allow users to add any number of rename operations
This adds a button that lets users choose a rename operation from a
dropdown. Doing this simplifies the UI and lets users choose the order
in which ops are applied.

Fixes Issue #39
Impacts Issue #21 which now simply requires reordering / deleting
rename operations.

For record keeping, here is the order I committed changed:
- Allow adding an arbitrary number of Rename Operations
- Allow RenameOps to specify their Display Name
- Add sorting to the Rename Ops in the context menu
- Revamp Window UI to use two columns
- Comply with StyleCop
- Fix scrolling on preview pane
- Cleanup function that took too many parameters