Compute Diff only after all RenameOps are applied
Instead of supplying an option to RenameOps to spit back out a string
that looks like a diff, I'm using a diff library. This will fix bugs
caused by passing a diff string into the next operation in a BulkRename.
I could still use the library to do an op-wise diff, but we'd need to
rework the GUI to show it.

Fixes Issue #50.