Cleanup and refactor of sprite renaming branch
- Remove custom sprite building in favor of AssetPreview

Rather than deal with the reimport of the texture and caching the
icon, we can just use AssetPreview.GetAssetPreview() for sprites.

- Clean up ObjectsToRename a bit

I considered adding a new class for ObjectsToRename, but it was very
nebulous and the only added functionality to a list was being auto-sorted
and validation for rename. In the end I decided it didn't really clean
it up at all.

- Move RenameAssets functions to BulkRenamer.cs

This cleans up separation of concerns a bit, so that the window doesn't
need to know about Renaming.

- Refactor Spritesheet renaming

This is a rework that builds on the concepts from the previous commit by
Uvivagabond. I moved all Renaming functionality into BulkRenamer as that
seems to be its responsibility, not BulkRenameWindow. I made a class to
better encapsulate renaming of spritesheets, which keeps all the hacky
meta file renaming in one place and lets us adapt it easily if Unity
changes. I also removed all the various methods for renaming.

- Fix Exception caused by AssetPreview resizing

The AssetPreview asynchronous load of the icon caused the preview row
to resize as the icons loaded and unloaded. I removed the null check
so that space is always reserved for the icon, whether it's loaded or
not.

- Improve display of progress on Rename

- Comply with StyleCop