Add support for undoing Asset renames
This required a decent rework in order to make the Undoer be able to
use BulkRenamer to reapply the renames. To detect undos on asset name
changes, we use a ScriptableObject that's added to the Undo stack,
and remains only in memory. The Undoer watches the Unity Undo callbacks
(which don't tell me whether or not it's a redo or undo) to determine
if it's an undo or redo, and uses a history of names to revert or
reapply undos.