Fix Issue #126 - Enumerate sometimes failing
To fix this I added a deferred step to the Rename function. If a renamed
file would collide with an existing file in the rename operation,
it is first named to a temporary value and added to a deferred list.
After all objects are renamed, the deferred list is then processed.