Use the same path to generate previews as renames
This cleans up an assumption that was made in the MulliganRenamerWindow
which used the RenameOperationSequence to infer the name change, instead
of going through BulkRenamer. This was basically an assumption that
BulkRenamer would also use the RenameOperationSequence to calculate its
results, but that may not always be the case. Now BulkRenamer has a
method to retrieve previews.

I also declare the RenameOperationSequence as a dependency in the
BulkRenamer constructor, instead of taking it as a parameter of the
rename function. Ultimately this may just be because it seemed weird to
have an object with zero state.