Cleanup minor issues with hardcoded sizes
Using a member variable seems like the least bad way to clean this up,
at this point. It really needs a refactor - probably to make a drawer
delegate or class for the preview row, which can be measured by
when trying to figure out how many rows can fit in the scroll view.