Add prompt to ask the user for a review (or purchase)
This adds preferences to track the number of times the tool is being
used and shows a prompt after it reaches a certain number. It also
adds a setting to RBPackageExporter to allow it to override a file
to specify the build came from GitHub and change the prompt accordingly.