Remove Positions side-effect from GetColoredTextsFromString
This removes the position calculation that was happening as part of the
GetColoredTextsFromString pass. I think it's cleaner to keep this
separate, and I try to avoid ref and out parameters if possible.

I also had to add a helper, GetConsecutiveBlocksOfToken() to find
blocks of either space or non-space within a colored text block,
in order to color only the background of whitespace.

I still don't have the measurements perfect, especially when adding in
blocks of characters and spaces (Ex: GameObject => replace 'e' with
"abc   edf   hij"). I'd like to figure out what's actually going on, but
for now this is workign ok.