Account for scroll position when resizing columns (Fixes Issue #223)
This was easy to fix when growing columns, but when shrinking them it
was tricky because shrinking the columns shrunk the scroll rect. This
caused it to scroll, which just infinitely shrunk columns until they
were minimally small. To address that I added a minimum width to content,
which is respected while resizing columns.