Initially size columns based on window size, not content
Now that you can double click the column to size to its contents, I
think it makes more sense to open the window with all columns visible.
Even if you can't see everything in a column, you get a sense of
what each column is. If you need to resize, you can.