Rework PreviewPanel to handle parameterized diff colors for #206
It was a challenge to make this able to take two different colors,
because BackgroundColor was calculated by using the color that was
stripped from the RichText tag of the previewed string.

In order to fix this, I went ahead and passed in the RenameResult
into the BackgroundColor function. This was going to make for cleaner
background coloring anyway.